move initConnection and handleAuth to ADB Deviceø
