+++ lib/usb-device.js
@@ -67,61 +67,6 @@ class USBDevice {
     this.inputEndpoint = Promise.promisifyAll(this.inputEndpoint);
   }
 
-  // handles authentication between our machine and
-  // the android device upon initial connection
-  // TODO: attempt to sign token sent to us rather than just send our .pub
-  async handleAuth (token) {
-    let publicKeyBuf = new Buffer(publicKeyString.length + 1);
-    publicKeyString.copy(publicKeyBuf);
-    publicKeyBuf[-1] = 0;
-
-    let signedToken = signLib.sign(new Buffer(privateKeyPath + "\0"), token);
-    // see if the device will accept our signed token
-    await this._sendMsg(ADB_COMMANDS.CMD_AUTH, 2, 0, signedToken);
-    let packet = await this._recvMsg(MAXDATA);
-    console.log("Device accepted our signed token? ", packet.command === ADB_COMMANDS.CMD_CNXN);
-    // if the signed token wasn't accepted then we'll send the public key
-    while (packet.command !== ADB_COMMANDS.CMD_CNXN) {
-      await this._sendMsg(ADB_COMMANDS.CMD_AUTH, 3 ,0, publicKeyBuf);
-      packet = await this._recvMsg(MAXDATA);
-      console.log("Sent auth message and public key.");
-        if (packet.command === ADB_COMMANDS.CMD_CNXN) {
-          // we should read the string about the device that gets sent back
-          return true;
-        }
-    }
-    return true;
-  }
-
-  // calls claim device and then performs the ADB usb handshake
-  // returns a promise, will have the error value if an error is thrown
-  async initConnection () {
-    await this.claimDevice();
-    console.log("Trying to connect to device.");
-    await this._sendMsg(ADB_COMMANDS.CMD_CNXN
-                       , CONNECT_VALUES.CONNECT_VERSION
-                       , CONNECT_VALUES.CONNECT_MAXDATA
-                       , CONNECT_VALUES.CONNECT_PAYLOAD);
-    console.log("Sent connect message.");
-    let packet = await this._recvMsg(MAXDATA);
-    if (packet.command === ADB_COMMANDS.CMD_AUTH) {
-      console.log("AUTH received.");
-      let run = true;
-      do {
-        let auth = await this.handleAuth(packet.data);
-        if (auth === true) {
-          run = false;
-        } else if (auth.errno === 2) { // timeout error
-          console.log("Read in handleAuth timed out");
-          continue;
-        } else {
-          throw auth;
-        }
-      } while (run === true);
-    }
-    console.log("End of initConnection.");
-  }
-
   // sends a message as an adb packet followed by the payload if necessary
   async _sendMsg (cmd, arg1, arg2, payload) {
     let adbPacket = generateMessage(cmd, arg1, arg2, payload);
